Update protocol.rst
I was reading through this in detail in order to design a PyTrilinos adaptor, and found some changes I wanted to make.

One question: a 'dim_data' tuple of length greater than one, where one or more of the dimension dictionaries has a 'dist_type' of 'u' would represent a structured array of unstructured dimensions.  I am not aware of any uses cases for this configuration.  Are you supporting it?  I would be OK with a constraint that if 'dist_type' == 'u', then len(dim_data) = 1 must be true.